# Enable tensor.pad lowering via buffer load with bounds check