@jerryyin 2025-03-26
I'm giving a summary of what has happened for the failing tiny matmul case:

So the problem is a matmul of shape `<2x4> of 2s` x `<4x2> of 3s` = `<2x2> of 24s`
 
For operand A:
 - thread 0 will load from [0, 0] with [4xtrue, 4xfalse]
 - thread 4 will load from [1, 0] with [4xtrue, 4xfalse]
 - All other threads will get [8xfalse]

For operand B:
 - thread 0 will load from [0, 0] with [2xtrue, 6xfalse]
 - thread 2 will load from [1, 0] with [2xtrue, 6xfalse]
 - thread 4 will load from [2, 0] with [2xtrue, 6xfalse]
 - **thread 6 will load from [3, 0] with [2xtrue, 6xfalse]**
 - All other threads will get [8xfalse]

As you can see, operand B of the last load from thread 6 will load from middle of the word and cross the boundary. So instead of loading value of vector size of 2 with 2s, it loads vector size of 2 with 0s because the word boundary crossed for the size. 

So we are instead computing:
`[[3, 3, 3, 3] [3, 3, 3, 3]]` x `[[2, 2], [2, 2], [2, 2], [0, 0]]` = `[[18, 18], [18, 18]]`
instead of the expected `[[24, 24], [24, 24]] `

----

This is now verified that @krzysz00's PR https://github.com/llvm/llvm-project/pull/132981 will address this issue. I'll rebase this PR once Krzysztof's PR is merged into upstream (and pulled into IREE)