@jerryyin 2025-03-26
Additional observations from @krzysz00's PR, is that the new PR now causes loads to de-vectorize because MLIR ops didn't carry alignment information.

Previously the LLVM IR we get:
```llvm
  %.loadable = call <2 x i32> @llvm.amdgcn.raw.ptr.buffer.load.v2i32(ptr addrspace(8) align 1 %4, i32 %35, i32 0, i32 0)
```

Now the new access pattern for both A and B operand both becomes:
```llvm
  %.off.0 = call i8 @llvm.amdgcn.raw.ptr.buffer.load.i8(ptr addrspace(8) align 1 %6, i32 %39, i32 0, i32 0)
  %.slice.0 = insertelement <8 x i8> poison, i8 %.off.0, i64 0
  %.off.ptr.1 = add nuw i32 %39, 1
  %.off.1 = call i8 @llvm.amdgcn.raw.ptr.buffer.load.i8(ptr addrspace(8) align 1 %6, i32 %.off.ptr.1, i32 0, i32 0)
  %.slice.1 = insertelement <8 x i8> %.slice.0, i8 %.off.1, i64 1
  %.off.ptr.2 = add nuw i32 %39, 2
  %.off.2 = call i8 @llvm.amdgcn.raw.ptr.buffer.load.i8(ptr addrspace(8) align 1 %6, i32 %.off.ptr.2, i32 0, i32 0)
  %.slice.2 = insertelement <8 x i8> %.slice.1, i8 %.off.2, i64 2
  %.off.ptr.3 = add nuw i32 %39, 3
  %.off.3 = call i8 @llvm.amdgcn.raw.ptr.buffer.load.i8(ptr addrspace(8) align 1 %6, i32 %.off.ptr.3, i32 0, i32 0)
  %.slice.3 = insertelement <8 x i8> %.slice.2, i8 %.off.3, i64 3
  %.off.ptr.4 = add nuw i32 %39, 4
  %.off.4 = call i8 @llvm.amdgcn.raw.ptr.buffer.load.i8(ptr addrspace(8) align 1 %6, i32 %.off.ptr.4, i32 0, i32 0)
  %.slice.4 = insertelement <8 x i8> %.slice.3, i8 %.off.4, i64 4
  %.off.ptr.5 = add nuw i32 %39, 5
  %.off.5 = call i8 @llvm.amdgcn.raw.ptr.buffer.load.i8(ptr addrspace(8) align 1 %6, i32 %.off.ptr.5, i32 0, i32 0)
  %.slice.5 = insertelement <8 x i8> %.slice.4, i8 %.off.5, i64 5
  %.off.ptr.6 = add nuw i32 %39, 6
  %.off.6 = call i8 @llvm.amdgcn.raw.ptr.buffer.load.i8(ptr addrspace(8) align 1 %6, i32 %.off.ptr.6, i32 0, i32 0)
  %.slice.6 = insertelement <8 x i8> %.slice.5, i8 %.off.6, i64 6
  %.off.ptr.7 = add nuw i32 %39, 7
```

@krzysz00 is investigating the approach for LLVM to infer alignment.