@jerryyin 2025-03-27
State of world: through CI results we learned that https://github.com/llvm/llvm-project/pull/132981 isn't the right approach because it is too pessimistic about the buffer load width (according to alignment). The damage is that for an in-bounds load with i8, #132981 will only load one byte at a time, killing the performance. In other words, #132981 will fix the correctness issue for boundary loads but kill performance of regular loads.

After discussion from the pre-sync meeting (https://github.com/iree-org/iree/issues/20392), the shortest path to address this is for @krzysz00 to implement masked load with buffer load according to alignment. This will allow `transfer_read` with a mask to load at reasonable width.

Once Krzysztof's new PR is ready I'll adopt his change as well as abandoning the `populateAmdgpuTransferReadToLoadPatterns`, so masked load get populated as-is from mlir to llvm, allowing the new implemented routine to take effect.