@jerryyin 2025-04-18
Note: reviewers: this PR is ready to review. The tip of our LLVM commit hash has caught up with upstream and all the correctness issues related with buffer load subword are fixed.